pyproject.toml
[project]
name = "animator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imageio[ffmpeg]>=2.37.0",
    "pillow>=11.3.0",
]
